Added hook-points
Now, Whisper has hook points. Therefore, you should ideally be able to
run with hooks. TODO(adriano) add unit tests to make sure it is doing
the same computation as the original (should be able to use pip
version).

There is a ipynb notebook so that you can test that the hooks work OK in
two simple cases.

Any sort of intermediate BS is removed.